==> Notebooks/01.EDA.ipynb <==
# Intro

- https://github.com/GoogleCloudDataproc/cloud-dataproc/tree/master/notebooks
- https://www.cloudskillsboost.google/focuses/3390?locale=fr&parent=catalog
- https://www.cloudskillsboost.google/focuses/5834?locale=fr&parent=catalog
- https://cloud.google.com/dataproc/docs/tutorials/gcs-connector-spark-tutorial?hl=fr#scala

spark:
- https://sparkbyexamples.com/spark/spark-sql-distinct-multiple-columns/
- https://mungingdata.com/apache-spark/dates-times/
- https://mungingdata.com/apache-spark/just-enough-scala/

println("Scala language: "+util.Properties.versionString)

import org.apache.spark.sql.SparkSession
import org.apache.spark.sql.expressions.Window
import java.time.LocalTime
import org.apache.spark.sql.functions.udf
import org.apache.spark.sql.DataFrame
import org.apache.spark.sql.functions._
import org.apache.spark.sql.functions.col
import org.apache.spark.sql.types._

// import spark.implicits._

import org.apache.spark.sql.SparkSession

val spark = SparkSession
  .builder()
  .appName("Spark SQL basic example")
  .config("spark.some.config.option", "some-value")
  .getOrCreate()

import spark.sqlContext.implicits._

// spark.sparkContext.version
spark.version

---

# Flights

val filePath = "gs://dataset-flight/Flights/201201.csv"
val df_flights = spark.read
  .option("header", "true")
  .option("inferSchema", "true")
  .csv(filePath)

df_flights.show(3)

val nbRows = df_flights.count()
print(nbRows)

df_flights.printSchema

// https://sparkbyexamples.com/spark/spark-find-count-of-null-empty-string-values/
// print(df_flights.filter(col("_c12").isNull || col("_c12") === "").count() / nbRows)

for (columnName <- df_flights.columns) println(columnName, df_flights.filter(col(columnName).isNull || col(columnName) === "").count() / nbRows)

for (columnName <- df_flights.columns) println("Nb nulls " + columnName + ": " + df_flights.filter(col(columnName).isNull || col(columnName) === "").count())

val flightColumns = List("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID", "DEST_AIRPORT_ID")

// df_flights.select(flightColumns.map(m=>col(m)):_*).describe().show()
df_flights.select(flightColumns.map(m=>col(m)):_*)
    .summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "max").show()

println("Distinct count date/carrier_id/fl_num: "+df_flights.select("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM").distinct().count())
println("Distinct count date/all iDs & origin : "+df_flights.select("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID").distinct().count())
println("Distinct count date/all iDs & numbers: "+df_flights.select("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID", "DEST_AIRPORT_ID").distinct().count())
println("Distinct count all columns           : "+df_flights.distinct().count())


identification de la clé de jointure

Le meme jour:
- Des vols de différentes company peuvent avoir le meme numéro le meme jour.  
- Des vols de la meme company peuvent avoir le meme numéro mais pas le meme aéroport de départ (heureusement !)

donc les champs suivants doivent servir de clés: "FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID"

val otherColumns = List("CRS_DEP_TIME", "ARR_DELAY_NEW", "CANCELLED", "DIVERTED", "CRS_ELAPSED_TIME", "WEATHER_DELAY", "NAS_DELAY")

df_flights.select(otherColumns.map(m=>col(m)):_*)
    .summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "max").show()

(2359/60, 1450.0/60, 1865.0/60 ,1179.0/60, 1007.0/60)

---
# Airport / WBAN / timezone

val airportsPath = "gs://dataset-flight/wban_airport_timezone.csv"
val df_airports = spark.read
  .option("header", "true")
  .option("inferSchema", "true")
  .csv(airportsPath)

df_airports.show(5)

df_airports.printSchema()

df_airports.summary("count", "count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "max").show()

---

# Weather

val weatherPath = "gs://dataset-flight/Weather/201201hourly.txt"
val df_weather = spark.read
  .option("header", "true")
  .option("inferSchema", "true")
  .csv(weatherPath)

df_weather.printSchema()

df_weather.select("WBAN", "Date", "Time", "StationType").show(3)

df_weather.select("WBAN", "Date", "Time", "StationType").summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "max").show()

a ne pas relancer car prend du temps:

for (columnName <- df_weather.columns) println("% nulls " + columnName + ": " + df_weather.filter(col(columnName).isNull || col(columnName) === "").count() / df_weather.count() * 100)

df_weather.columns

https://www.ncei.noaa.gov/resources/metadata
Quality Controlled Local Climatological Data (QCLCD

df_weather.select(df_weather.columns.take(12).map(col): _*).show(10)

for (columnName <- df_weather.columns.filter(x => x.endsWith("Flag"))) {
  println("10 of the disctinct values for " + columnName + ": " + df_weather.select(columnName).distinct().take(10).mkString(", "))
}

for (columnName <- df_weather.columns.filter(x => !x.endsWith("Flag"))) {
  println("10 of the disctinct values for " + columnName + ": " + df_weather.select(columnName).distinct().take(10).mkString(", "))
}

une grande partie des features sont en string à cause de l'espace à la place des NULL ! recalcul du % du null par colonne:

val nbRows = df_weather.count()
for (columnName <- df_weather.columns) {
    println("% nulls " + columnName + ": " + df_weather.filter(col(columnName).isNull || col(columnName) === "" || col(columnName) === " " || col(columnName) === "  ").count() * 100 / nbRows)
}
