==> Notebooks/02.Preprocessing.ipynb <==
# Intro

https://stackoverflow.com/questions/39926411/provide-schema-while-reading-csv-file-as-a-dataframe-in-scala-spark

import org.apache.spark.sql.SparkSession
import org.apache.spark.sql.expressions.Window
import java.time.LocalTime
import org.apache.spark.sql.functions.udf
import org.apache.spark.sql.DataFrame
import org.apache.spark.sql.functions._
import org.apache.spark.sql.functions.col
import org.apache.spark.sql.types._

// import spark.implicits._

import org.apache.spark.sql.SparkSession

val spark = SparkSession
  .builder()
  .appName("Spark SQL basic example")
  .config("spark.some.config.option", "some-value")
  .getOrCreate()

import spark.sqlContext.implicits._

println("Scala language: "+util.Properties.versionString)

// spark.sparkContext.version
spark.version

---

# Flights

val filePath = "gs://dataset-flight/Flights/"
val df_flights_raw = spark.read
  .option("header", "true")
  .option("inferSchema", "true")
  .csv(filePath)

println(df_flights_raw.count())
df_flights_raw.show(3)

486133 -> 18286055 pour toutes les csv

preprocessing

val delayThreshold = 30

val df_flights_clean = df_flights_raw
    .drop("_c12") // 478k
    .where(col("CANCELLED") === 0).drop("CANCELLED")
    .where(col("DIVERTED") === 0).drop("DIVERTED") // 478k
    .na.fill(0, Seq("NAS_DELAY"))
    .withColumn("NAS_DELAY", col("NAS_DELAY").cast("int"))
    .dropDuplicates(Seq("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID")) // 437k
    .withColumn("CRS_ELAPSED_TIME", col("CRS_ELAPSED_TIME").cast("int")) // float -> int
    .withColumn("ARR_DELAY_NEW", col("ARR_DELAY_NEW").cast("int"))
    .where(!(col("ARR_DELAY_NEW") < -45)) // n'arrive jamais 45min en avance
    .where(col("ARR_DELAY_NEW") < (5.3 * 60)) // 437k
    .withColumn("IS_DELAYED", when(col("ARR_DELAY_NEW") >= delayThreshold and (col("NAS_DELAY") >= delayThreshold or col("WEATHER_DELAY") > 0), 1).otherwise(0))

println(df_flights_clean.count())

df_flights_clean.show(3)

df_flights_clean.select("IS_DELAYED")
    .summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "95%", "max").show()

df_flights_raw.select("WEATHER_DELAY")
    .summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "95%", "max").show()

df_flights_clean.printSchema()

df_flights_clean.printSchema()

nas_delay = la grande majorité à NULL sinon 0:

df_flights_raw.select("NAS_DELAY")
    .summary("count", "approx_count_distinct", "mean", "stddev", "min", "25%", "50%", "75%", "max").show()

---

une fois les vols filtrés (cancelled, diverted...), statistiques en heures:

val df_temp = df_flights_raw
    .drop("_c12") // 478k
    .where(col("CANCELLED") === 0).drop("CANCELLED")
    .where(col("DIVERTED") === 0).drop("DIVERTED") // 478k
    .where(col("NAS_DELAY").isNull || col("NAS_DELAY") === 0).drop("NAS_DELAY") // 437k
    .dropDuplicates(Seq("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID")) // 437k
    .withColumn("CRS_ELAPSED_TIME", col("CRS_ELAPSED_TIME").cast("int")) // float -> int
    .withColumn("ARR_DELAY_NEW", col("ARR_DELAY_NEW").cast("int"))
    .where(!(col("ARR_DELAY_NEW") < -45)) // n'arrive jamais 45min en avance

df_temp
    .select("CRS_DEP_TIME", "CRS_ELAPSED_TIME", "ARR_DELAY_NEW", "WEATHER_DELAY")
    .summary("mean", "stddev", "min", "25%", "50%", "75%", "90%", "95%", "max")
    .withColumn("CRS_DEP_TIME", round(col("CRS_DEP_TIME") / 60, 1))
    .withColumn("CRS_ELAPSED_TIME", round(col("CRS_ELAPSED_TIME") / 60, 1))
    .withColumn("ARR_DELAY_NEW", round(col("ARR_DELAY_NEW") / 60, 1))
    .withColumn("WEATHER_DELAY", round(col("WEATHER_DELAY") / 60, 1))
    .show()

- 25% des vols sont programmés entre 00H00 et 15H00
- 25% des vols sont programmés entre 15H00 et 22H00
- 25% des vols sont programmés entre 22H00 et 04H00 du matin le lendemain

concernant la durée de vols:
- 5% des vols ont une durée prévisionelle de plus de 5h, le maximum étant une durée prévisionnelle de 11h.

pour ce qui est des retards:
- 75% des vols sont a l'heure (ARR_DELAY à zero)
- la moyenne des retard est de 6min (0.1h)
- 90% des vols ont moins de 12 min de retard

concernant la ponctualité au départ ou a l'arrivée: celles ci sont de l'ordre de 80%:
https://www.lefigaro.fr/voyages/conseils/ponctualite-des-avions-les-aeroports-francais-sont-toujours-a-la-traine-20230728

sachant que la durée des vols ne peut pas dépasser de plus de 3 ou 4h (pour les vols courts) et 5 ou 6h (pour les vols longs) pour des questions de réserve de carburant limitées, les ARR_DELAY_NEW de plus de 5h doivent forcément soit être du à des départs repoussés ou des valeurs erronées. Or nous ne connaissons pas de combien de temps le décollage a été repoussés, mais ils y en a forcément vu que des CRS_DEP_TIME dépassent les 24h ! __Supprimons les arrival delay et weather delay trop importants de respectivement plus de 2.4 et 2.2h. --> OUTLIERS __

df_temp
    .select("ARR_DELAY_NEW", "WEATHER_DELAY")
    .summary("90%", "95%", "97%", "99%", "99.5%", "99.95%", "max")
    .withColumn("ARR_DELAY_NEW", round(col("ARR_DELAY_NEW") / 60, 1))
    .withColumn("WEATHER_DELAY", round(col("WEATHER_DELAY") / 60, 1))
    .show()

---

# Weather

val weatherPath = "gs://dataset-flight/Weather/201201hourly.txt"

val weatherSchema = StructType(Array(    
    StructField("WBAN", IntegerType, true),
    StructField("Date", IntegerType, true),
    StructField("Time", IntegerType, true),
    StructField("StationType", IntegerType, true),
    StructField("SkyCondition", StringType, true),
    StructField("SkyConditionFlag", StringType, true),
    StructField("Visibility", StringType, true), // enlever espace puis caster en int
    StructField("VisibilityFlag", StringType, true),
    StructField("WeatherType", StringType, true),
    StructField("WeatherTypeFlag", StringType, true),
    StructField("DryBulbFarenheit", IntegerType, true),
    StructField("DryBulbFarenheitFlag", StringType, true),
    StructField("DryBulbCelsius", FloatType, true),
    StructField("DryBulbCelsiusFlag", StringType, true),
    StructField("WetBulbFarenheit", IntegerType, true),
    StructField("WetBulbFarenheitFlag", StringType, true),
    StructField("WetBulbCelsius", FloatType, true),
    StructField("WetBulbCelsiusFlag", StringType, true),
    StructField("DewPointFarenheit", IntegerType, true),
    StructField("DewPointFarenheitFlag", StringType, true),
    StructField("DewPointCelsius", FloatType, true),
    StructField("DewPointCelsiusFlag", StringType, true),
    StructField("RelativeHumidity", StringType, true), // a convertir en int après space
    StructField("RelativeHumidityFlag", StringType, true),
    StructField("WindSpeed", IntegerType, true),
    StructField("WindSpeedFlag", StringType, true),
    StructField("WindDirection", IntegerType, true),
    StructField("WindDirectionFlag", StringType, true),
    StructField("ValueForWindCharacter", IntegerType, true),
    StructField("ValueForWindCharacterFlag", StringType, true),
    StructField("StationPressure", FloatType, true),
    StructField("StationPressureFlag", StringType, true),
    StructField("PressureTendency", IntegerType, true),
    StructField("PressureTendencyFlag", StringType, true),
    StructField("PressureChange", IntegerType, true),
    StructField("PressureChangeFlag", StringType, true),
    StructField("SeaLevelPressure", FloatType, true),
    StructField("SeaLevelPressureFlag", StringType, true),
    StructField("RecordType", StringType, true),
    StructField("RecordTypeFlag", StringType, true),
    StructField("HourlyPrecip", FloatType, true),
    StructField("HourlyPrecipFlag", StringType, true),
    StructField("Altimeter", FloatType, true),
    StructField("AltimeterFlag", StringType, true),
))


val df_weather_raw = spark.read.format("csv")
    .option("header", "true")
    .option("delimited", ",")
    .schema(weatherSchema)
    .option("treatEmptyValuesAsNulls","true")
    .option("nullValue", null)
    .option("emptyValue", null)
    .load(weatherPath)


println(df_weather_raw.count())
println(df_weather_raw.columns.size)
// df_weather_raw.printSchema()

    // .drop("_c12") // 478k
    // .where(col("CANCELLED") === 0).drop("CANCELLED")
    // .where(col("DIVERTED") === 0).drop("DIVERTED") // 478k
    // .where(col("NAS_DELAY").isNull || col("NAS_DELAY") === 0).drop("NAS_DELAY") // 437k
    // .dropDuplicates(Seq("FL_DATE", "OP_CARRIER_AIRLINE_ID", "OP_CARRIER_FL_NUM", "ORIGIN_AIRPORT_ID")) // 437k
    // .withColumn("CRS_ELAPSED_TIME", col("CRS_ELAPSED_TIME").cast("int")) // float -> int
    // .withColumn("ARR_DELAY_NEW", col("ARR_DELAY_NEW").cast("int"))
    // .where(!(col("ARR_DELAY_NEW") < -45)) // n'arrive jamais 45min en avance
    // .where(col("ARR_DELAY_NEW") < (5.3 * 60)) // 437k
    // // .where(col("WEATHER_DELAY") < (4.4 * 60)) // supprimer trop de ligne bizarrement ??

val df_weather_clean = df_weather_raw
    .drop(df_weather_raw.columns.filter(colName => colName.endsWith("Flag")) : _*)
    .withColumn("Visibility", trim(col("Visibility")).cast("int"))
    .withColumn("RelativeHumidity", trim(col("RelativeHumidity")).cast("int"))


println(df_weather_clean.count())
println(df_weather_clean.columns.size)

println("Visibility: "+df_weather_clean.select("Visibility").distinct().take(10).mkString(", "))
println("RelativeHumidity: "+df_weather_clean.select("RelativeHumidity").distinct().take(10).mkString(", "))

// for (columnName <- df_weather_raw.columns) {
//   println("10 of the disctinct values for " + columnName + ": " + df_weather_raw.select(columnName).distinct().take(10).mkString(", "))
// }

for (columnName <- df_weather_clean.columns) {
  println("10 of the disctinct values for " + columnName + ": " + df_weather_clean.select(columnName).distinct().take(10).mkString(", "))
}

val nbRows = df_weather_clean.count()

for (columnName <- df_weather_clean.columns) {
    println("% nulls " + columnName + ": " + (df_weather_clean.filter(col(columnName).isNull || col(columnName) === "" || col(columnName) === " " || col(columnName) === "null").count() * 100.0 / nbRows)
}

val nbRows = df_weather_clean.count()

for (columnName <- df_weather_clean.columns) {
    println("% nulls " + columnName + ": " + (df_weather_clean.filter(col(columnName).isNull || col(columnName) === "" || col(columnName) === " " || col(columnName) === "null").count() * 100.0) / nbRows)
}

---

# Backup

import spark.sqlContext.implicits._
import org.apache.spark.sql.functions._
val test = Seq(("2019-01-23"),("2019-06-24"),("2019-09-20")).toDF("date")

test.show()

import java.sql.Timestamp

val df = Seq(
    ("notebook",    Timestamp.valueOf("2019-01-29 12:00:00"), 2),
    ("notebook",    Timestamp.valueOf("2019-01-01 00:00:00"), 3),
    ("small_phone", Timestamp.valueOf("2019-01-15 23:00:00"), 4),
    ("small_phone", Timestamp.valueOf("2019-01-01 09:00:00"), 5)
).toDF("device", "purchase_time", "minutes").sort("device","purchase_time")

df.show()

df.withColumn("new", to_timestamp(col("purchase_time"))).show()

val testbis = test.withColumn("date", to_timestamp(col("date"), "yyyy-MM-dd")) // HH:mm:ss
testbis.show()

testbis.printSchema()

import org.apache.spark.sql.functions.{add_months, date_add, _}


testbis.select(col("date"), add_months(col("date"), 3).as("add_months")).show()

import spark.sqlContext.implicits._
import org.apache.spark.sql.functions._

testbis.add_months(lit(3)).show()


df.select(col("date"),
    add_months(col("date"),3).as("add_months"), // provide +ve value to add months
    add_months(col("date"),-3).as("sub_months"), //provide -ve value to subtract months
    date_add(col("date"),4).as("date_add"), // to add day
    date_sub(col("date"),4).as("date_sub") //to substract day
  ).show()

import org.apache.spark.sql.functions._

df_flights_clean
    .withColumn("FL_DATE_TIME", col("FL_DATE")  + (col("CRS_DEP_TIME") * 60))
// df = df.withColumn("timestamp", F.expr("from_unixtime(unix_timestamp(concat_ws(' ', date, time)) + (`additional_time(in mins)` * 60))"))

    .show()

spark.sql("select input_timestamp, " +
    "cast(input_timestamp as TIMESTAMP) + INTERVAL 2 hours as added_hours," +
    "cast(input_timestamp as TIMESTAMP) + INTERVAL 5 minutes as added_minutes," +
    "cast(input_timestamp as TIMESTAMP) + INTERVAL 55 seconds as added_seconds from AddTimeExample"
    )
