==> Notebooks/05.ML_nonull_stratified_undersample_RF_SELECtIOIN.ipynb <==
import java.time.LocalTime
import org.apache.spark.sql.SparkSession

import org.apache.spark.sql.SparkSession
import org.apache.spark.sql.expressions.Window
import org.apache.spark.sql.functions.udf
import org.apache.spark.sql.DataFrame
import org.apache.spark.sql.functions._
import org.apache.spark.sql.functions.col
import org.apache.spark.sql.types._
import org.apache.spark.sql.functions.lag
import org.apache.spark.sql.expressions.Window

import org.apache.spark.mllib.tree.model.DecisionTreeModel
import org.apache.spark.mllib.tree.DecisionTree
import org.apache.spark.mllib.util.MLUtils
import org.apache.spark.mllib.evaluation.MulticlassMetrics 
import org.apache.spark.mllib.evaluation.BinaryClassificationMetrics  

import org.apache.spark.ml.Pipeline
import org.apache.spark.ml.classification.{RandomForestClassificationModel, RandomForestClassifier}
import org.apache.spark.ml.evaluation.MulticlassClassificationEvaluator
import org.apache.spark.ml.feature.{IndexToString, StringIndexer, VectorIndexer, VectorAssembler}


val spark = SparkSession
  .builder()
  .appName("Spark ML")
  .config("spark.some.config.option", "some-value")
  .getOrCreate()

import spark.sqlContext.implicits._

val bucket = "dataproc-temp-us-central1-1044206227610-i54vpwyj"
spark.conf.set("temporaryGcsBucket", bucket)

println("Scala language: "+util.Properties.versionString)

// spark.sparkContext.version
spark.version

val data_raw = spark.read.parquet("gs://dataset-flight/final-for-ml.parquet")
//     .drop("ARR_DATETIME_RND")
//     .drop("ARR_WBAN")

val data = data_raw
    .drop(data_raw.columns.filter(colName => colName.contains("2")) : _*)
    .drop(data_raw.columns.filter(colName => colName.contains("3")) : _*)
    .drop(data_raw.columns.filter(colName => colName.contains("WindSpeed")) : _*)
    .na.drop()

val nbRows = data.count()
println(data.columns.size, nbRows)

data.select("IS_DELAYED").summary("count", "mean").show()

// columns that need to added to feature column
val featColumns = data.columns.filter(name => !name.contains("IS_DELAYED"))
println(data.columns.size, featColumns.length)

// VectorAssembler to add feature column
// input columns - cols
// feature column - features
val vectorAssembler = new VectorAssembler()
    .setInputCols(featColumns)
    .setOutputCol("features")
    .setHandleInvalid("skip") // options are "keep", "error" or "skip"


val featureDf = vectorAssembler.transform(data)
// featureDf.printSchema()


// StringIndexer define new 'label' column with 'result' column
// Index labels, adding metadata to the label column.
// Fit on whole dataset to include all labels in index.
val labelIndexer = new StringIndexer()
  .setInputCol("IS_DELAYED")
  .setOutputCol("label")


val labelDf = labelIndexer.fit(featureDf).transform(featureDf)
// labelDf.printSchema()

// Split the data into training and test sets (30% held out for testing).
// split data set training and test
// training data set - 70%
// test data set - 30%
val seed = 42

val zeros = labelDf.filter(col("IS_DELAYED") === 0).sample(0.1, seed)
val ones = labelDf.filter(col("IS_DELAYED") === 1)

// split datasets into training and testing
val Array(train0, test0) = zeros.randomSplit(Array(0.7, 0.3), seed)
val Array(train1, test1) = ones.randomSplit(Array(0.7, 0.3), seed)

// stack datasets back together
val trainingData = train0.union(train1)
val testData = test0.union(test1)

println(f"""There are ${trainingData.count} rows in the training set, and ${testData.count} in the test set""")

testData.select("IS_DELAYED").summary("count", "mean").show()

trainingData.select("IS_DELAYED").summary("count", "mean").show()

// train Random Forest model with training data set
val randomForestClassifier = new RandomForestClassifier()
  .setImpurity("gini")
  .setMaxDepth(15)
  .setNumTrees(30)
  .setFeatureSubsetStrategy("auto")
  .setSeed(seed)
  .setLabelCol("label") // indexedLabel
  .setFeaturesCol("features") // indexedFeatures

val randomForestModel = randomForestClassifier.fit(trainingData)

val predictionDf = randomForestModel.transform(testData)

// Select example rows to display.
predictionDf.select("IS_DELAYED", "features", "label", "rawPrediction", "probability", "prediction").show(5)

Confusion matrix

val predictionAndLabels = predictionDf
    .select("prediction", "label")
    .as[(Double, Double)]
    .rdd

// Instantiate a new metrics objects
val bMetrics = new BinaryClassificationMetrics(predictionAndLabels)
val mMetrics = new MulticlassMetrics(predictionAndLabels)
val labels = mMetrics.labels

// Print out the Confusion matrix
println("Confusion matrix:")
println(mMetrics.confusionMatrix)

val trueNegative = predictionDf.filter(col("prediction") === 0 && col("label") === col("prediction")).count().toDouble
val truePositive = predictionDf.filter(col("prediction") === 1 && col("label") === col("prediction")).count().toDouble
val falseNegative = predictionDf.filter(col("prediction") === 0 && col("label") =!= col("prediction")).count().toDouble
val falsePositive = predictionDf.filter(col("prediction") === 1 && col("label") =!= col("prediction")).count().toDouble

val totalDelays = predictionDf.filter(col("label") === 1).count()

val precision = truePositive / (truePositive + falsePositive) 
val recall = truePositive / (truePositive + falseNegative) 
val acc = (truePositive + trueNegative) / (truePositive + trueNegative + falsePositive + falseNegative) 
val f1score = 2 * precision * recall / (precision + recall) 

## Conclusion:
- nulls intelligement + FEAT selection on enlève _2 & _3
- stratified split + undersampling
- Random Forest MaxDepth(__15__)  .setNumTrees(__30__)
- precision = 0.75
- recall = 0.24
- acc = 0.81
- f1score = 0.36

feature importance

https://stackoverflow.com/questions/47043836/scala-random-forest-feature-importance-extraction-with-names-labels

val featureImportance = decisionTreeModel.featureImportances

// featureImportance.getClass

val res = featColumns.zip(featureImportance.toArray).sortBy(-_._2)


for (e <- res.take(20)) println(e)
